Created a mockup file
We need to have a calendar to show events in the future. If every entry is created like this one, is there a way to say that only dates in the future get posted, an maybe even a maximum amount of three future events with the possibility to expand to more?